browser_playwright/pyproject.toml
[project]
name = "browser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bedrock-agentcore>=0.1.3",
    "playwright>=1.55.0",
    "strands-agents>=1.8.0",
]
